fix "RuntimeError: "bernoulli_tensor_cpu_self_" not implemented for 'Half'" (fix #130)

+++ scripts/model_mixer.py
@@ -3606,7 +3606,7 @@ Direct Download: <a href="{s['downloadUrl']}" target="_blank">{s["filename"]} [{
             # Generate the mask m^t from Bernoulli distribution
             #m = torch.from_numpy(np.random.binomial(1, p, theta0.shape)).to(tensor1.dtype) # slow
             if "GPU" in calc_settings:
-                m = torch.bernoulli(torch.full_like(input=theta0, fill_value=p)).to(device="cuda")
+                m = torch.bernoulli(torch.full_like(input=theta0.float(), fill_value=p)).to(device="cuda")
             else:
                 m = torch.bernoulli(torch.full_like(input=theta0.float(), fill_value=p))
             # Apply the mask to the delta to get δ̃^t
@@ -3618,7 +3618,7 @@ Direct Download: <a href="{s['downloadUrl']}" target="_blank">{s["filename"]} [{
             # alpha = alpha / (1 - p) * m
             alpha = torch.mul(m, alpha / (1 - p))
             if "GPU" in calc_settings:
-                return torch.lerp(theta0, theta1, alpha).to(theta0.dtype).cpu()
+                return torch.lerp(theta0.float(), theta1.float(), alpha.float().cpu()).to(theta0.dtype)
 
             return torch.lerp(theta0.float(), theta1.float(), alpha).to(theta0.dtype)
